Visuart/backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
dnspython==2.4.1
firebase-admin==6.2.0
Flask==2.3.2
Flask-PyMongo==2.3.0
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
openai==0.27.8
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.4.1
pyparsing==3.1.1
requests==2.31.0
rsa==4.9
six==1.16.0
tqdm==4.65.1
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2

opencv-python~=4.8.1.78
numpy~=1.24.3
tensorflow~=2.13.0
Pillow~=10.0.0